// engemedical-worker/src/mongo/enum/scheduling.enum.ts
export enum ExamStatus {
  PENDENTE = 'PENDENTE',
  AGUARDANDO_RESULTADO = 'AGUARDANDO_RESULTADO',
  FINALIZADO = 'FINALIZADO',
}

export enum AtendimentoStatus {
  AGENDADO = 'AGENDADO',
  EM_ATENDIMENTO = 'ATENDIMENTO',
  AGUARDANDO_RESULTADOS = 'AGUARDANDO_RESULTADOS',
  AGUARDANDO_AVALIACAO_MEDICA = 'AVALIACAO_MEDICA',
  FINALIZADO = 'FINALIZADO',
}

export enum AsoStatus {
  NAO_GERADO = 'NAO_GERADO',
  GERADO = 'GERADO',
  KIT_CREDENCIADA = 'KIT_CREDENCIADA',
  APROVADO = 'APROVADO',
  REPROVADO = 'REPROVADO',
  ENVIADO = 'ENVIADO',
}

export enum MongoOperationTypes {
  INSERT = 'insert',
  REPLACE = 'replace',
  UPDATE = 'update',
  DELETE = 'delete',
}

export enum ParecerMedico {
  APTO = 'APTO',
  APTO_COM_ORIENTACAO = 'APTO_COM_ORIENTACAO',
  APTO_COM_RESTRICAO = 'APTO_COM_RESTRICAO',
  SOLICITAR_REPETICAO = 'SOLICITAR_REPETICAO',
  INAPTO = 'INAPTO',
  INAPTO_TEMPORARIAMENTE = 'INAPTO_TEMPORARIAMENTE',
}